RavenDB-13111 address PR comments

+++ src/Raven.Client/Json/BlittableArrayResult.cs
@@ -5,7 +5,7 @@ namespace Raven.Client.Json
     public class BlittableArrayResult
     {
         public BlittableJsonReaderArray Results { get; set; }
-        public int TotalResults { get; set; }
+        public long TotalResults { get; set; }
     }
 
     public class BatchCommandResult
